pub_sub/requirements_dev.txt
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
contourpy==1.1.1
cycler==0.12.1
fonttools==4.53.1
google-api-core==2.19.1
google-auth==2.33.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-pubsub==2.23.0
google-cloud-storage==2.18.1
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpc-google-iam-v1==0.13.1
grpcio==1.65.4
grpcio-status==1.65.4
idna==3.7
importlib_resources==6.4.0
kiwisolver==1.4.5
matplotlib==3.7.5
numpy==1.24.4
packaging==24.1
pandas==2.0.3
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
rsa==4.9
seaborn==0.13.2
six==1.16.0
tzdata==2024.1
urllib3==2.2.2
zipp==3.19.2
